requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==2.0.41
Flask-Cors==4.0.0
Flask-Login==0.6.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.31.0
